backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.0
pytest==8.3.3
httpx==0.27.2
groq==0.11.0
resend==0.7.0
supabase==2.5.1
